backend/requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
flask-cors==4.0.1
supabase==2.5.1
python-dotenv==1.0.1
